feature010: Added validations, added image and description to user groups
