+++ formulapanadera/src/app/modules/settings/user-groups/listing/user-groups-listing.page.html
@@ -36,7 +36,7 @@
       *ngFor="let user_group of user_groups; index as i"
       [user_group]="user_group"
       [even]="(i%2==0)"
-      (click)="presentOptions(user_group)"
+      (click)="updateUserGroup(user_group)"
     ></app-user-group-item>
   </ion-list>
   <ng-template #emptyBlock>
