+++ formulapanadera/src/app/modules/settings/user-groups/manage/styles/user-groups-manage.page.scss
@@ -111,6 +111,10 @@ ion-back-button {
   padding: 1px;
 }
 
+.colored-field {
+  color: var(--ion-color-primary);
+}
+
 .no-padding {
   padding: 0;
 }
