+++ formulapanadera/src/app/modules/settings/user-groups/manage/user-groups-manage.page.ts
@@ -1,7 +1,7 @@
 import { Component, NgZone, OnInit } from "@angular/core";
 import { FormControl, FormGroup, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
-import { LoadingController, ToastController } from '@ionic/angular';
+import { ActionSheetController, AlertController, LoadingController, ToastController } from '@ionic/angular';
 import { APP_URL } from 'src/app/config/configuration';
 import { ICONS } from 'src/app/config/icons';
 import { UserGroupModel, UserModel, UserResumeModel } from 'src/app/core/models/user.model';
@@ -12,7 +12,7 @@ import { UserStorageService } from 'src/app/core/services/storage/user.service';
 @Component({
   selector: "app-user-groups-manage",
   templateUrl: "./user-groups-manage.page.html",
-  styleUrls: ["./styles/user-groups-manage.page.scss"],
+  styleUrls: ["./styles/user-groups-manage.page.scss", "../../../../shared/styles/confirm.alert.scss"],
 })
 export class UserGroupsManagePage implements OnInit {
   ICONS = ICONS;
@@ -32,6 +32,8 @@ export class UserGroupsManagePage implements OnInit {
     private languageService: LanguageService,
     private loadingController: LoadingController,
     private toastController: ToastController,
+    private actionSheetController: ActionSheetController,
+    private alertController: AlertController,
     private ngZone: NgZone
   ){}
 
@@ -41,13 +43,19 @@ export class UserGroupsManagePage implements OnInit {
     if (state == undefined) {
       this.manageUserGroupForm = new FormGroup({
         name: new FormControl(null, Validators.required),
+        description: new FormControl(null, null),
+        image_url: new FormControl(null, null),
       });
     } else {
       this.update = true;
       this.manageUserGroupForm = new FormGroup({
         name: new FormControl(state.user_group.name, Validators.required),
+        description: new FormControl(state.user_group.description, null),
+        image_url: new FormControl(state.user_group.image_url, null),
       });
-      this.user_group.name = state.user_group.name
+      this.user_group.name = state.user_group.name;
+      this.user_group.description = state.user_group.description;
+      this.user_group.image_url = state.user_group.image_url;
       state.user_group.users.forEach((user) => {
         this.user_group.users.push(user);
       });
@@ -62,42 +70,59 @@ export class UserGroupsManagePage implements OnInit {
     });
     await loading.present();
 
-    if (this.update) {
-      this.user.user_groups.forEach((group, group_index) => {
-        if (group.name == this.user_group.name) {
-          this.user.user_groups[group_index].name = this.manageUserGroupForm.value.name
-          this.user.user_groups[group_index].users = this.user_group.users
+    let group_exists = false;
+    if (this.user.user_groups && this.user.user_groups.length>0) {
+      this.user.user_groups.forEach((group) => {
+        if (group.name == this.manageUserGroupForm.value.name) {
+          group_exists = true;
         }
-      })
-    } else {
-      this.user_group.name = this.manageUserGroupForm.value.name;
-      this.user.user_groups.push(this.user_group)
-    }
-    this.userCRUDService
-      .updateUser(this.user)
-      .then(async () => {
-        await this.userStorageService
-          .setUser(this.user)
-          .then(() => {
-            this.ngZone.run(() =>
-              this.router.navigate(
-                [
-                  APP_URL.menu.name +
-                  "/" +
-                  APP_URL.menu.routes.settings.main +
-                  "/" +
-                  APP_URL.menu.routes.settings.routes.user_groups.main
-                ]
-              )
-            );
-          });
-      })
-      .catch(() => {
-        this.presentToast(false);
-      })
-      .finally(async () => {
-        await loading.dismiss();
       });
+    }
+    if (group_exists) {
+      loading.dismiss();
+      this.presentToast(false, true);
+    } else {
+      if (this.update) {
+        this.user.user_groups.forEach((group, group_index) => {
+          if (group.name == this.user_group.name) {
+            this.user.user_groups[group_index].name = this.manageUserGroupForm.value.name;
+            this.user.user_groups[group_index].description = this.manageUserGroupForm.value.description;
+            this.user.user_groups[group_index].image_url = this.manageUserGroupForm.value.image_url;
+            this.user.user_groups[group_index].users = this.user_group.users;
+          }
+        })
+      } else {
+        this.user_group.name = this.manageUserGroupForm.value.name;
+        this.user_group.description = this.manageUserGroupForm.value.description;
+        this.user_group.image_url = this.manageUserGroupForm.value.image_url;
+        this.user.user_groups.push(this.user_group)
+      }
+      this.userCRUDService
+        .updateUser(this.user)
+        .then(async () => {
+          await this.userStorageService
+            .setUser(this.user)
+            .then(() => {
+              this.ngZone.run(() =>
+                this.router.navigate(
+                  [
+                    APP_URL.menu.name +
+                    "/" +
+                    APP_URL.menu.routes.settings.main +
+                    "/" +
+                    APP_URL.menu.routes.settings.routes.user_groups.main
+                  ]
+                )
+              );
+            });
+        })
+        .catch(() => {
+          this.presentToast(false);
+        })
+        .finally(async () => {
+          await loading.dismiss();
+        });
+    }
   }
 
   addUser() {
@@ -136,9 +161,11 @@ export class UserGroupsManagePage implements OnInit {
     return valid;
   }
 
-  async presentToast(success: boolean) {
+  async presentToast(success: boolean, exists: boolean = false) {
     const toast = await this.toastController.create({
-      message: success
+      message: exists
+        ? this.languageService.getTerm("send.exists")
+        : success
         ? this.languageService.getTerm("send.success")
         : this.languageService.getTerm("send.error"),
       color: "secondary",
@@ -153,5 +180,89 @@ export class UserGroupsManagePage implements OnInit {
       ],
     });
     toast.present();
+  }
+
+  async presentOptions(user_group: UserGroupModel) {
+    const actionSheet = await this.actionSheetController.create({
+      cssClass: "formula-options",
+      buttons: [
+        {
+        text: this.languageService.getTerm("action.delete"),
+        icon: ICONS.trash,
+        cssClass: "delete-icon",
+          handler: () => {
+            this.deleteUserGroup(user_group);
+          }
+        },
+        {
+          text: this.languageService.getTerm("action.cancel"),
+          icon: ICONS.close,
+          role: "cancel",
+          cssClass: "cancel-icon",
+          handler: () => {},
+        },
+      ],
+    });
+    await actionSheet.present();
+  }
+
+  async deleteUserGroup(user_group: UserGroupModel) {
+    let user = await this.userStorageService.getUser();
+    user.user_groups.forEach((group, index) => {
+      if (group.name == user_group.name) {
+        user.user_groups.splice(index, 1)
+      }
+    })
+    const alert = await this.alertController.create({
+      header: this.languageService.getTerm("action.confirm"),
+      message: this.languageService.getTerm("action.delete_question", {
+        item: user_group.name,
+      }),
+      cssClass: "confirm-alert",
+      buttons: [
+        {
+          text: this.languageService.getTerm("action.cancel"),
+          role: "cancel",
+          handler: () => {},
+        },
+        {
+          text: this.languageService.getTerm("action.ok"),
+          cssClass: "confirm-alert-accept",
+          handler: async () => {
+            const loading = await this.loadingController.create({
+              cssClass: "app-send-loading",
+              message: this.languageService.getTerm("loading"),
+            });
+            await loading.present();
+
+            this.userCRUDService
+              .updateUser(user)
+              .then(() => {
+                this.userStorageService.setUser(user)
+                  .then(() => {
+                    this.ngZone.run(() =>
+                      this.router.navigate(
+                        [
+                          APP_URL.menu.name +
+                          "/" +
+                          APP_URL.menu.routes.settings.main +
+                          "/" +
+                          APP_URL.menu.routes.settings.routes.user_groups.main
+                        ]
+                      )
+                    );
+                  })
+              })
+              .catch(() => {
+                this.presentToast(false);
+              })
+              .finally(async () => {
+                await loading.dismiss();
+              });
+          },
+        },
+      ],
+    });
+    await alert.present();
   }
 }
