+++ formulapanadera/src/app/shared/components/user-group/user-group-item/user-group-item.component.ts
@@ -1,4 +1,5 @@
 import { Component, Input } from "@angular/core";
+import { ASSETS } from "src/app/config/assets";
 import { APP_URL } from 'src/app/config/configuration';
 import { ICONS } from 'src/app/config/icons';
 import { UserGroupModel } from 'src/app/core/models/user.model';
@@ -14,6 +15,7 @@ import { UserGroupModel } from 'src/app/core/models/user.model';
 export class UserGroupItemComponent {
   ICONS = ICONS
   APP_URL = APP_URL
+  DEFAULT_ICON = ASSETS.icons.default
 
   @Input() user_group: UserGroupModel;
   @Input() even: boolean = false;
