+++ formulapanadera/src/assets/i18n/en.json
@@ -181,6 +181,8 @@
     "user_groups": {
       "name": "User groups",
       "user_group_name": "Name",
+      "user_group_description": "Description",
+      "user_group_image_url": "Image (URL)",
       "users": "Users",
       "create": "Create user group",
       "empty": "No user groups found"
