+++ formulapanadera/src/assets/i18n/es.json
@@ -181,6 +181,8 @@
     "user_groups": {
       "name": "Grupos de usuarios",
       "user_group_name": "Nombre",
+      "user_group_description": "Descripción",
+      "user_group_image_url": "Imagen (URL)",
       "users": "Usuarios",
       "create": "Crear grupo",
       "empty": "No se encontraron grupos de usuarios"
